// src/app/types/adminGeneralPages.ts
export interface ContactUsData {
  id?: string;
  email: string;
  phone: number;
  managerEmail: string;
  resumeEmail: string;
  address: string;
}

export interface AboutUsData {
  id?: string;
  title: string;
  description: string;
}

export interface ShoppingGuideProps {
  id?: string;
  title: string;
  description: string;
}